+++ src/Common/EngineCore/Events/EventIds.cs
@@ -266,25 +266,19 @@ public enum EventId
     ///     Removes a global key-value pair.
     /// </summary>
     /// Associated details: StringEventDetails (specifying the key)
-    Core_Data_RemoveGlobal = 901,
-
-    /// <summary>
-    ///     Announces that a global key-value pair has been created.
-    /// </summary>
-    /// Associated details: KeyValueEventDetails
-    Core_Data_GlobalAdded = 902,
+    Core_Data_RemoveGlobal = 902,
 
     /// <summary>
     ///     Announces that a global key-value pair has been updated.
     /// </summary>
     /// Associated details: KeyValueEventDetails
-    Core_Data_GlobalUpdated = 902,
+    Core_Data_GlobalSet = 903,
 
     /// <summary>
     ///     Announces that a global key-value pair has been removed.
     /// </summary>
     /// Associated details: StringEventDetails (specifying the key)
-    Core_Data_GlobalRemoved = 903,
+    Core_Data_GlobalRemoved = 904,
 
     #endregion Data