+++ src/Server/Persistence/PersistenceServiceCollectionExtensions.cs
@@ -18,6 +18,7 @@ using Microsoft.Extensions.DependencyInjection;
 using Microsoft.Extensions.DependencyInjection.Extensions;
 using Sovereign.EngineCore.Systems;
 using Sovereign.Persistence.Accounts;
+using Sovereign.Persistence.Data;
 using Sovereign.Persistence.Database;
 using Sovereign.Persistence.Entities;
 using Sovereign.Persistence.Players;
@@ -40,6 +41,7 @@ public static class PersistenceServiceCollectionExtensions
     public static IServiceCollection AddSovereignPersistence(this IServiceCollection services)
     {
         AddAccounts(services);
+        AddData(services);
         AddDatabase(services);
         AddEntities(services);
         AddPlayers(services);
@@ -52,6 +54,11 @@ public static class PersistenceServiceCollectionExtensions
     private static void AddAccounts(IServiceCollection services)
     {
         services.TryAddSingleton<PersistenceAccountServices>();
+    }
+
+    private static void AddData(IServiceCollection services)
+    {
+        services.TryAddSingleton<GlobalKeyValueProcessor>();
     }
 
     private static void AddDatabase(IServiceCollection services)